requirements.txt
defusedxml~=0.7.1
cffi~=1.15.0
Pillow~=9.0.1
pip~=21.2.4
tornado~=6.1
amqp~=5.0.9
vine~=5.0.0
pytz~=2021.3
Django~=4.0.2
kombu~=5.2.3
billiard~=3.6.4.0
redis~=4.1.3
requests~=2.27.1
Deprecated~=1.2.13
packaging~=21.3
celery~=5.2.3
click~=8.0.3
Pygments~=2.11.2
sqlparse~=0.4.2
asgiref~=3.5.0
colorama~=0.4.4
humanize~=4.0.0
multidict~=6.0.2
httpie~=3.0.2
urllib3~=1.26.8
certifi~=2021.10.8
six~=1.16.0
wcwidth~=0.2.5
idna~=3.3
pyparsing~=3.0.7
pycparser~=2.21
wrapt~=1.13.3
setuptools~=59.6.0
djangorestframework~=3.13.1